+++ paddle/phi/kernels/cpu/randint_kernel.cc
@@ -23,8 +23,8 @@ namespace phi {
 
 template <typename T, typename Context>
 void RandintRawKernel(const Context& dev_ctx,
-                      int low,
-                      int high,
+                      int64_t low,
+                      int64_t high,
                       const ScalarArray& shape,
                       DataType dtype,
                       int seed,
@@ -47,8 +47,8 @@ void RandintRawKernel(const Context& dev_ctx,
 
 template <typename T, typename Context>
 void RandintKernel(const Context& dev_ctx,
-                   int low,
-                   int high,
+                   int64_t low,
+                   int64_t high,
                    const ScalarArray& shape,
                    DataType dtype,
                    DenseTensor* out) {
